pyproject.toml
[project]
name = "k8s-home"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "packaging>=26.0",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "ruamel-yaml>=0.19.1",
]
